backend/requirements.txt
fastapi
uvicorn
motor
pydantic[email]
passlib[bcrypt]
python-multipart
python-dotenv
PyJWT
